+++ incl/commands/NickCommand.hpp
