+++ refs/allowedFunctions.txt
